pyproject.toml
[project]
name = "events-face"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django>=5.2.3",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.0",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
